New WebUIDataManager class
This commit adds a new WebUIDataManager class that inherit Shinken
DataManager and half of the methods that where provided by the main
Webui_broker class.

Most of the modifications consist in replacing "app.method()" by
"app.datamgr.method()".

I also cleanup or remove some datamanager methods, and make
datamanager.py PEP8 compliant.

Related to #178
