+++ module/plugins/dashboard/views/dashboard.tpl
@@ -15,8 +15,8 @@
             <a href="/all?search=type:host isnot:OK" class="btn btn-sm">
                <i class="fa fa-4x fa-server font-darkgrey"></i>
                <span class="badger-title hosts"><i class="fa fa-plus" style="color: #ccc"></i>&nbsp;Hosts</span>
-               %host_state = app.get_percentage_hosts_state(app.get_user_auth(), False)
-               <span class="badger-big badger-left alert-inverse">{{app.get_nb_hosts(app.get_user_auth())}}</span>
+               %host_state = app.datamgr.get_percentage_hosts_state(app.get_user_auth(), False)
+               <span class="badger-big badger-left alert-inverse">{{len(app.datamgr.get_hosts(app.get_user_auth()))}}</span>
                <span class="badger-big badger-right alert-{{'critical' if host_state <= app.hosts_states_warning else 'warning' if host_state <= app.hosts_states_critical else 'ok'}}">{{host_state}}%</span>
             </a>
          </td>
@@ -25,8 +25,8 @@
             <a href="/all?search=type:service isnot:OK" class="btn btn-sm">
                <i class="fa fa-4x fa-bars font-darkgrey"></i>
                <span class="badger-title services"><i class="fa fa-plus" style="color: #ccc"></i>&nbsp;Services</span>
-               %service_state = app.get_percentage_service_state(app.get_user_auth(), False)
-               <span class="badger-big badger-left alert-inverse">{{app.get_nb_services(app.get_user_auth())}}</span>
+               %service_state = app.datamgr.get_percentage_service_state(app.get_user_auth(), False)
+               <span class="badger-big badger-left alert-inverse">{{len(app.datamgr.get_services(app.get_user_auth()))}}</span>
                <span class="badger-big badger-right alert-{{'critical' if service_state <= app.services_states_warning else 'warning' if service_state <= app.services_states_critical else 'ok'}}">{{service_state}}%</span>
             </a>
          </td>
@@ -41,8 +41,8 @@
             <a href="/problems" class="btn btn-sm">
                <i class="fa fa-4x fa-exclamation-triangle font-darkgrey"></i>
                <span class="badger-title itproblem"><i class="fa fa-plus" style="color: #ccc"></i>&nbsp;IT Problems</span>
-               %overall_itproblem = app.get_overall_it_state(app.get_user_auth())
-               <span title="Number of not acknowledged IT problems." class="badger-big alert-{{'ok' if overall_itproblem == 0 else 'warning' if overall_itproblem == 1 else 'critical'}}">{{app.get_overall_it_problems_count(user, get_acknowledged=False)}}</span>
+               %overall_itproblem = app.datamgr.get_overall_it_state(app.get_user_auth())
+               <span title="Number of not acknowledged IT problems." class="badger-big alert-{{'ok' if overall_itproblem == 0 else 'warning' if overall_itproblem == 1 else 'critical'}}">{{app.datamgr.get_overall_it_problems_count(user, get_acknowledged=False)}}</span>
             </a>
          </td>
 
@@ -50,8 +50,8 @@
             <a href="/impacts" class="btn btn-sm">
                <i class="fa fa-4x fa-flash font-darkgrey"></i>
                <span class="badger-title impacts"><i class="fa fa-plus" style="color: #ccc"></i>&nbsp;Impacts</span>
-               %overall_state = app.get_overall_state(app.get_user_auth())
-               <span title="Number of not acknownledged IT problems." class="badger-big alert-{{'ok' if overall_state == 0 else 'warning' if overall_state == 1 else 'critical'}}">{{app.get_overall_state_problems_count(user)}}</span>
+               %overall_state = app.datamgr.get_overall_state(app.get_user_auth())
+               <span title="Number of not acknownledged IT problems." class="badger-big alert-{{'ok' if overall_state == 0 else 'warning' if overall_state == 1 else 'critical'}}">{{app.datamgr.get_overall_state_problems_count(user)}}</span>
             </a>
          </td>
       </tr>
