+++ module/plugins/groups/views/hosts-groups-dashboard.tpl
@@ -1,7 +1,7 @@
 %rebase("layout", css=['groups/css/groups-overview.css'], title='Hosts groups dashboard', refresh=True)
 
 %helper = app.helper
-%hosts = app.get_hosts(user)
+%hosts = app.datamgr.get_hosts(user)
 %h = helper.get_synthesis(hosts)['hosts']
 
 
@@ -78,7 +78,7 @@
       <tbody>
       %i=0
       %for group in hostgroups:
-         %hosts = app.search_hosts_and_services('type:host hg:'+group.get_name(), user, hosts_only=True)
+         %hosts = app.datamgr.search_hosts_and_services('type:host hg:'+group.get_name(), user, hosts_only=True)
          %h = helper.get_synthesis(hosts)['hosts']
          %nHosts=h['nb_elts']
          %nGroups=len(group.get_hostgroup_members())
