+++ module/plugins/tags/tags.py
@@ -42,8 +42,8 @@ def show_tags():
     user = app.check_user_authentication()
 
     fake_htags = []
-    for tag in app.get_host_tags_sorted():
-        hosts = only_related_to(app.get_hosts_tagged_with(tag[0]), user)
+    for tag in app.datamgr.get_host_tags_sorted():
+        hosts = only_related_to(app.datamgr.get_hosts_tagged_with(tag[0]), user)
         if len(hosts) > 0:
             fake_htags.append({'name': tag[0], 'hosts': hosts})
 
@@ -54,8 +54,8 @@ def show_stags():
     user = app.check_user_authentication()
 
     fake_stags = []
-    for tag in app.get_service_tags_sorted():
-        services = only_related_to(app.get_services_tagged_with(tag[0]), user)
+    for tag in app.datamgr.get_service_tags_sorted():
+        services = only_related_to(app.datamgr.get_services_tagged_with(tag[0]), user)
         if len(services) > 0:
             fake_stags.append({'name': tag[0], 'services': services})
