+++ module/plugins/timeperiods/timeperiods.py
@@ -37,7 +37,7 @@ params['display'] = ''
 def load_cfg():
     global params
     
-    import os,sys
+    import os
     from webui.config_parser import config_parser
     from shinken.log import logger
     plugin_name = os.path.splitext(os.path.basename(__file__))[0]
